+++ otterdog/webapp/webhook/github_models.py
@@ -1,5 +1,5 @@
 #  *******************************************************************************
-#  Copyright (c) 2023-2024 Eclipse Foundation and others.
+#  Copyright (c) 2023-2025 Eclipse Foundation and others.
 #  This program and the accompanying materials are made available
 #  under the terms of the Eclipse Public License 2.0
 #  which is available at http://www.eclipse.org/legal/epl-v20.html
@@ -163,6 +163,7 @@ class Issue(BaseModel):
 class WorkflowJob(BaseModel):
     name: str
     id: int
+    workflow_name: str | None
     run_id: int
     runner_id: int | None = None
     runner_name: str | None = None
@@ -174,6 +175,26 @@ class WorkflowJob(BaseModel):
     completed_at: str | None = None
     conclusion: str | None = None
     labels: list[str]
+
+
+class WorkflowRef(BaseModel):
+    path: str
+    ref: str | None
+    sha: str
+
+
+class WorkflowRun(BaseModel):
+    name: str
+    id: int
+    run_number: int
+    run_attempt: int
+    status: str
+    head_sha: str
+    head_branch: str | None = None
+    created_at: str
+    run_started_at: str
+    conclusion: str | None = None
+    referenced_workflows: list[WorkflowRef] | None
 
 
 class Commit(BaseModel):
@@ -244,6 +265,14 @@ class WorkflowJobEvent(Event):
 
     action: str
     workflow_job: WorkflowJob
+    repository: Repository
+
+
+class WorkflowRunEvent(Event):
+    """A payload sent for workflow run events."""
+
+    action: str
+    workflow_run: WorkflowRun
     repository: Repository